pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saisa"
version = "2.0.0"
description = "SAISA - Super AI Self-Autonomous: a powerful terminal-based coding agent inspired by Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Souraka HAMIDA", email = "hamidasouraka@gmail.com" },
]
keywords = ["ai", "agent", "coding", "terminal", "llm", "ollama", "groq", "openai", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
groq = ["openai>=1.0.0"]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.20.0"]
search = ["duckduckgo_search>=6.0"]
browser = ["playwright>=1.40.0"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "duckduckgo_search>=6.0",
    "playwright>=1.40.0",
]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
]

[project.scripts]
saisa = "saisa.cli:main"

[project.urls]
Homepage = "https://github.com/Souraka229/local-agent-windows"
Repository = "https://github.com/Souraka229/local-agent-windows"

[tool.hatch.build.targets.wheel]
packages = ["saisa"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
